+++ mzpick-front/src/MzPick.tsx
@@ -13,9 +13,9 @@ import SignUp from './views/Auth/SignUp';
 import Fashion from './views/Fashion';
 import Food from './views/Food';
 import HOF from './views/HOF';
-import HOFFashion from './views/HOF/HOFFashion';
-import HOFFood from './views/HOF/HOFFood';
-import HOFTravel from './views/HOF/HOFTravel';
+import HOFFashion from './views/HOF/FashionHof';
+import HOFFood from './views/HOF/FoodHof';
+import HOFTravel from './views/HOF/TravelHof';
 import Home from './views/Home';
 import Keyword from './views/Keyword';
 import MyPage from './views/MyPage';
