+++ mzpick-front/src/views/HOF/FoodHof/index.tsx
@@ -1,22 +1,22 @@
 
 import { useNavigate } from 'react-router';
-import { HOF_FASHION_PATH, HOF_FOOD_PATH, HOF_TRAVEL_PATH } from '../../constants';
-// import './style.css';
+import { HOF_FASHION_PATH, HOF_FOOD_PATH, HOF_TRAVEL_PATH } from 'src/constants';
+import './style.css';
 
-function TravelTop1() {  
+function FoodTop1() {
 
-  // function: 네비게이션 상태 //
-  const navigator = useNavigate();
+    // function: 네비게이션 상태 //
+    const navigator = useNavigate();
 
+    // event handler: 명예의전당 buttonbox 클릭 이벤트 //
+    const onButtonClickEventHandler = (path: string) =>{
+      navigator(path);
+    };
 
-  // event handler: 명예의전당 buttonbox 클릭 이벤트 //
-  const onButtonClickEventHandler = (path: string) =>{
-    navigator(path);
-  };
   return (
     <div id='main-hof-wrapper'>
       <div className='item-box'>
-        <div className='text-area'>"Travel Top 1"</div>
+        <div className='text-area'>"Food Top 1"</div>
         <div className='icon-area'></div>
       </div>
 
@@ -25,24 +25,23 @@ function TravelTop1() {
       
       <div className='deco-container'>
       <div className="decorated-img"></div>
-      </div>
       <div className='button-contaner'>
       <div className='button-box' onClick={() => onButtonClickEventHandler(HOF_TRAVEL_PATH)}>○</div>
       <div className='button-box' onClick={() => onButtonClickEventHandler(HOF_FOOD_PATH)}>○</div>
       <div className='button-box'onClick={() => onButtonClickEventHandler(HOF_FASHION_PATH)}>○</div>
       </div>
-
+      </div>
       </div>
     </div>
   )
 }
 
-export default function HOFTravel() {
+export default function HOFFood() {
 
 
   return (
     <>
-      <TravelTop1 />
+      <FoodTop1 />
     </>
   )
 }
