--- mzpick-front/src/views/HOF/HOFFashion.tsx
@@ -1,6 +1,0 @@
-
-export default function HOFFashion() {
-    return (
-        <div>Fashionindex</div>
-    )
-}
