--- mzpick-front/src/views/HOF/HOFFood.tsx
@@ -1,6 +1,0 @@
-
-export default function HOFFood() {
-    return (
-        <div>Foodindex</div>
-    )
-}
