--- mzpick-front/src/views/HOF/HOFTravel.tsx
@@ -1,21 +1,0 @@
-
-function TravelTop1(){
-    return (
-        <div id='hof-main-wrapper'>
-            <div className="title-hof">"TRAVEL TOP 1"</div>
-            <div className="icon-hof"></div>
-            <div className="image-wrapper">
-            <div className="image-section"></div>
-            <div className="decorated-section"></div>
-            </div>
-        </div>
-    )
-}
-
-export default function HOFTravel() {
-
-    return (
-        <div><TravelTop1 />
-        </div>
-    )
-}
