+++ mzpick-front/src/views/HOF/TravelHof/style.css
@@ -1,0 +1,76 @@
+#main-hof-wrapper {
+    display: flex;
+    flex-direction: column;
+    width: 90%;
+    height: 90%;
+    right: 15%;
+}
+
+
+#main-hof-wrapper .item-box {
+    display: flex;
+    position: relative;
+    justify-content: space-between;
+    margin-top: 54px;
+    left: 10%;
+}
+
+#main-hof-wrapper .item-box .text-area {
+    font-family: 'Pretendard-Regular';
+    src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff') format('woff');
+    font-weight: bold;
+    font-size: 40px;
+}
+
+#main-hof-wrapper .item-box .icon-area {
+    background-image: url(/public/image/medal.png);
+    width: 400px;
+    height: 400px;
+    object-fit: scale-down;
+}
+
+#main-hof-wrapper .main-contents {
+    display: flex;
+    position: absolute;
+    top: 1%;
+    left: 15%;
+    bottom: 15%;
+    justify-content: center;
+    align-items: center;
+    width: 70vw;
+    height: 100vh;
+    /* background-color: black; */
+}
+
+#main-hof-wrapper .main-contents .image-area {
+    position: absolute;
+    justify-content: center;
+    align-items: center;
+    border: solid 1px;
+    border-radius: 100%;
+    width: 480px;
+    height: 480px;
+}
+
+#main-hof-wrapper .deco-container {
+    width: 100%;
+    height: auto;
+}
+
+#main-hof-wrapper .deco-container .decorated-img {
+    position: relative;
+    left: 18.3%;
+    z-index: -1;
+    width: 649px;
+    height: 620px;
+    background-image: url(/public/image/deco.png);
+}
+
+#main-hof-wrapper .button-contaner{
+    display: flex;
+    flex-direction: row;
+    gap: 8px;
+    justify-content: center;
+    left: 1%;
+    cursor: pointer;
+}
