+++ mzpick-front/src/views/HOF/index.tsx
@@ -1,6 +1,13 @@
+import HOFFashion from './FashionHof'
+import HOFFood from './FoodHof'
+import HOFTravel from './TravelHof'
 
 export default function HOF() {
     return (
-        <div>index</div>
+        <>
+        <HOFTravel/>
+        <HOFFood/>
+        <HOFFashion/>
+        </>
     )
 }
